refactor(fs): modularize models and recommendation logic
